requirements.txt
streamlit
Pillow
folium
streamlit-folium
pandas
numpy
opencv-python